+++ UI/src/app/app-routing.module.ts
@@ -2,14 +2,12 @@ import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from './auth/auth.guard';
 
+
 const routes: Routes = [
   { path: '', redirectTo: 'posts', pathMatch: 'full' },
   { path: 'auth', loadChildren: './auth/auth.module#AuthPageModule', canLoad: [AuthGuard] },
-  {
-    path: 'posts',
-    loadChildren: './posts/posts.module#PostsPageModule',
-    canLoad: [AuthGuard]
-  },
+  { path: 'bookings', loadChildren: './bookings/bookings.module#BookingsPageModule', canLoad:[AuthGuard] },
+  { path: 'posts', loadChildren: './posts/posts.module#PostsPageModule', canLoad: [AuthGuard]},
   {
     path: 'user',
     children: [
@@ -24,8 +22,7 @@ const routes: Routes = [
         canLoad: [AuthGuard]
       }
     ]
-  },
-  { path: 'bookings', loadChildren: './bookings/bookings.module#BookingsPageModule', canLoad:[AuthGuard] }
+  }
 ];
 
 @NgModule({
