--- UI/src/app/bookings/booking.model.ts
@@ -1,10 +1,0 @@
-
-
-export class Booking {
-
-    constructor(
-        public id: string,
-        public postId: string,
-        public userId: string
-    ) {}
-}
