--- UI/src/app/posts/post.model.ts
@@ -1,15 +1,0 @@
-
-
-export class Post {
-    constructor(
-        public id: string,
-        public title: string,
-        public des: string,
-        public imgUrl: string,
-        public userId: string,
-        public date: Date,
-        public loc: string = 'Rabat'
-    ) {}
-}
-
-
