+++ UI/src/app/user/edit-user/edit-user.page.html
@@ -3,12 +3,15 @@
     <ion-buttons slot="start">
       <ion-back-button [defaultHref]="['/', 'user', userId]"></ion-back-button>
     </ion-buttons>
+
     <ion-title>edit-user</ion-title>
+    
     <ion-buttons slot="primary">
       <ion-button (click)="onEditUser()" [disabled]="!form.valid">
         <ion-icon name="checkmark" slot="icon-only"></ion-icon>
       </ion-button>
     </ion-buttons>
+  
   </ion-toolbar>
 </ion-header>
