+++ UI/src/app/user/user.page.html
@@ -3,12 +3,15 @@
     <ion-buttons slot="start">
       <ion-menu-button></ion-menu-button>
     </ion-buttons>
+
     <ion-title>user</ion-title>
+    
     <ion-buttons slot="primary">
       <ion-button (click)="onEditUser()">
         <ion-icon name="create" slot="icon-only"></ion-icon>
       </ion-button>
     </ion-buttons>
+  
   </ion-toolbar>
 </ion-header>
